Fix a bug on too large value in Memory consumption introducing long type

+++ src/Patch/VectorPatch.cpp
@@ -971,27 +971,27 @@ void VectorPatch::move_probes(Params& params, double x_moved)
 // Print information on the memory consumption
 void VectorPatch::check_memory_consumption(SmileiMPI* smpi)
 {
-    int particlesMem(0);
+    long int particlesMem(0);
     for (unsigned int ipatch=0 ; ipatch<size() ; ipatch++)
         for (unsigned int ispec=0 ; ispec<patches_[ipatch]->vecSpecies.size(); ispec++)
             particlesMem += patches_[ipatch]->vecSpecies[ispec]->getMemFootPrint();
     MESSAGE( 1, "(Master) Species part = " << (int)( (double)particlesMem / 1024./1024.) << " Mo" );
     
-    double dParticlesMem = (double)particlesMem / 1024./1024./1024.;
-    MPI_Reduce( smpi->isMaster()?MPI_IN_PLACE:&dParticlesMem, &dParticlesMem, 1, MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD );
+    long double dParticlesMem = (double)particlesMem / 1024./1024./1024.;
+    MPI_Reduce( smpi->isMaster()?MPI_IN_PLACE:&dParticlesMem, &dParticlesMem, 1, MPI_LONG_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD );
     MESSAGE( 1, setprecision(3) << "Global Species part = " << dParticlesMem << " Go" );
     
     MPI_Reduce( smpi->isMaster()?MPI_IN_PLACE:&particlesMem, &particlesMem, 1, MPI_INT, MPI_MAX, 0, MPI_COMM_WORLD );
     MESSAGE( 1, "Max Species part = " << (int)( (double)particlesMem / 1024./1024.) << " Mb" );
     
     // fieldsMem contains field per species
-    int fieldsMem(0);
+    long int fieldsMem(0);
     for (unsigned int ipatch=0 ; ipatch<size() ; ipatch++)
         fieldsMem += patches_[ipatch]->EMfields->getMemFootPrint();
     MESSAGE( 1, "(Master) Fields part = " << (int)( (double)fieldsMem / 1024./1024.) << " Mo" );
     
-    double dFieldsMem = (double)fieldsMem / 1024./1024./1024.;
-    MPI_Reduce( smpi->isMaster()?MPI_IN_PLACE:&dFieldsMem, &dFieldsMem, 1, MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD );
+    long double dFieldsMem = (double)fieldsMem / 1024./1024./1024.;
+    MPI_Reduce( smpi->isMaster()?MPI_IN_PLACE:&dFieldsMem, &dFieldsMem, 1, MPI_LONG_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD );
     MESSAGE( 1, setprecision(3) << "Global Fields part = " << dFieldsMem << " Go" );
     
     MPI_Reduce( smpi->isMaster()?MPI_IN_PLACE:&fieldsMem, &fieldsMem, 1, MPI_INT, MPI_MAX, 0, MPI_COMM_WORLD );
